IsMissedOut deleted / iban and price columns moved
